requirements.txt
fastapi
uvicorn
httpx
groq
gspread
google-auth
python-dotenv